factor out EntityService
