+++ core/src/main/scala/org/broadinstitute/dsde/rawls/entities/EntityService.scala
@@ -1,0 +1,270 @@
+package org.broadinstitute.dsde.rawls.entities
+
+import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport._
+import akka.http.scaladsl.model.{StatusCodes, Uri}
+import com.typesafe.scalalogging.LazyLogging
+import org.broadinstitute.dsde.rawls.dataaccess.slick.{DataAccess, ReadWriteAction}
+import org.broadinstitute.dsde.rawls.dataaccess.{SamDAO, SlickDataSource, SlickWorkspaceContext}
+import org.broadinstitute.dsde.rawls.expressions.ExpressionEvaluator
+import org.broadinstitute.dsde.rawls.metrics.RawlsInstrumented
+import org.broadinstitute.dsde.rawls.model.AttributeUpdateOperations.{AttributeUpdateOperation, EntityUpdateDefinition}
+import org.broadinstitute.dsde.rawls.model.WorkspaceJsonSupport._
+import org.broadinstitute.dsde.rawls.model.{AttributeEntityReference, Entity, EntityCopyDefinition, EntityQuery, EntityQueryResponse, EntityQueryResultMetadata, ErrorReport, SamResourceTypeNames, SamWorkspaceActions, UserInfo, WorkspaceName, _}
+import org.broadinstitute.dsde.rawls.util.{JsonFilterUtils, WorkspaceSupport}
+import org.broadinstitute.dsde.rawls.webservice.PerRequest
+import org.broadinstitute.dsde.rawls.webservice.PerRequest.{PerRequestMessage, RequestComplete}
+import org.broadinstitute.dsde.rawls.workspace.AttributeUpdateOperationException
+import org.broadinstitute.dsde.rawls.{RawlsException, RawlsExceptionWithErrorReport}
+import spray.json.DefaultJsonProtocol._
+
+import scala.concurrent.{ExecutionContext, Future}
+import scala.util.{Failure, Success, Try}
+
+object EntityService {
+  def constructor(dataSource: SlickDataSource, samDAO: SamDAO, workbenchMetricBaseName: String)
+                 (userInfo: UserInfo)
+                 (implicit executionContext: ExecutionContext) = {
+
+    // create the EntityManager along with its associated provider-builders. Since entities are only accessed
+    // in the context of a workspace, this is safe/correct to do here. We also want to use the same dataSource
+    // and execution context for the rawls entity provider that the workspace service uses.
+    val defaultEntityProviderBuilder = new RawlsEntityProviderBuilder(dataSource) // implicit executionContext
+    // soon: create a provider-builder for data repo and add it to the entity manager
+    val entityManager = new EntityManager(Set(defaultEntityProviderBuilder))
+
+    new EntityService(userInfo, dataSource, samDAO, entityManager, workbenchMetricBaseName)
+  }
+}
+
+class EntityService(protected val userInfo: UserInfo, val dataSource: SlickDataSource, val samDAO: SamDAO, entityManager: EntityManager, override val workbenchMetricBaseName: String)(implicit protected val executionContext: ExecutionContext)
+  extends WorkspaceSupport with LazyLogging with RawlsInstrumented with JsonFilterUtils {
+
+  import dataSource.dataAccess.driver.api._
+
+  def CreateEntity(workspaceName: WorkspaceName, entity: Entity) = createEntity(workspaceName, entity)
+  def GetEntity(workspaceName: WorkspaceName, entityType: String, entityName: String) = getEntity(workspaceName, entityType, entityName)
+  def UpdateEntity(workspaceName: WorkspaceName, entityType: String, entityName: String, operations: Seq[AttributeUpdateOperation]) = updateEntity(workspaceName, entityType, entityName, operations)
+  def DeleteEntities(workspaceName: WorkspaceName, entities: Seq[AttributeEntityReference]) = deleteEntities(workspaceName, entities)
+  def RenameEntity(workspaceName: WorkspaceName, entityType: String, entityName: String, newName: String) = renameEntity(workspaceName, entityType, entityName, newName)
+  def EvaluateExpression(workspaceName: WorkspaceName, entityType: String, entityName: String, expression: String) = evaluateExpression(workspaceName, entityType, entityName, expression)
+  def GetEntityTypeMetadata(workspaceName: WorkspaceName) = entityTypeMetadata(workspaceName)
+  def ListEntities(workspaceName: WorkspaceName, entityType: String) = listEntities(workspaceName, entityType)
+  def QueryEntities(workspaceName: WorkspaceName, entityType: String, query: EntityQuery) = queryEntities(workspaceName, entityType, query)
+  def CopyEntities(entityCopyDefinition: EntityCopyDefinition, uri:Uri, linkExistingEntities: Boolean) = copyEntities(entityCopyDefinition, uri, linkExistingEntities)
+  def BatchUpsertEntities(workspaceName: WorkspaceName, entityUpdates: Seq[EntityUpdateDefinition]) = batchUpdateEntities(workspaceName, entityUpdates, true)
+  def BatchUpdateEntities(workspaceName: WorkspaceName, entityUpdates: Seq[EntityUpdateDefinition]) = batchUpdateEntities(workspaceName, entityUpdates, false)
+
+
+  def createEntity(workspaceName: WorkspaceName, entity: Entity): Future[Entity] =
+    withAttributeNamespaceCheck(entity) {
+      getWorkspaceContextAndPermissions(workspaceName, SamWorkspaceActions.write) flatMap { workspaceContext =>
+        entityManager.createEntity(workspaceContext.workspace, entity)
+      }
+    }
+
+  def getEntity(workspaceName: WorkspaceName, entityType: String, entityName: String): Future[PerRequestMessage] =
+    getWorkspaceContextAndPermissions(workspaceName, SamWorkspaceActions.read) flatMap { workspaceContext =>
+      dataSource.inTransaction { dataAccess =>
+        withEntity(workspaceContext, entityType, entityName, dataAccess) { entity =>
+          DBIO.successful(PerRequest.RequestComplete(StatusCodes.OK, entity))
+        }
+      }
+    }
+
+  def updateEntity(workspaceName: WorkspaceName, entityType: String, entityName: String, operations: Seq[AttributeUpdateOperation]): Future[PerRequestMessage] =
+    withAttributeNamespaceCheck(operations.map(_.name)) {
+      getWorkspaceContextAndPermissions(workspaceName, SamWorkspaceActions.write) flatMap { workspaceContext =>
+        dataSource.inTransaction { dataAccess =>
+          withEntity(workspaceContext, entityType, entityName, dataAccess) { entity =>
+            val updateAction = Try {
+              val updatedEntity = applyOperationsToEntity(entity, operations)
+              dataAccess.entityQuery.save(workspaceContext, updatedEntity)
+            } match {
+              case Success(result) => result
+              case Failure(e: AttributeUpdateOperationException) =>
+                DBIO.failed(new RawlsExceptionWithErrorReport(errorReport = ErrorReport(StatusCodes.BadRequest, s"Unable to update entity ${entityType}/${entityName} in ${workspaceName}", ErrorReport(e))))
+              case Failure(regrets) => DBIO.failed(regrets)
+            }
+            updateAction.map(RequestComplete(StatusCodes.OK, _))
+          }
+        }
+      }
+    }
+
+  def deleteEntities(workspaceName: WorkspaceName, entRefs: Seq[AttributeEntityReference]): Future[PerRequestMessage] =
+    getWorkspaceContextAndPermissions(workspaceName, SamWorkspaceActions.write) flatMap { workspaceContext =>
+      dataSource.inTransaction { dataAccess =>
+        withAllEntities(workspaceContext, dataAccess, entRefs) { entities =>
+          dataAccess.entityQuery.getAllReferringEntities(workspaceContext, entRefs.toSet) flatMap { referringEntities =>
+            if (referringEntities != entRefs.toSet)
+              DBIO.successful(RequestComplete(StatusCodes.Conflict, referringEntities))
+            else {
+              dataAccess.entityQuery.hide(workspaceContext, entRefs).map(_ => RequestComplete(StatusCodes.NoContent))
+            }
+          }
+        }
+      }
+    }
+
+  def renameEntity(workspaceName: WorkspaceName, entityType: String, entityName: String, newName: String): Future[PerRequestMessage] =
+    getWorkspaceContextAndPermissions(workspaceName, SamWorkspaceActions.write) flatMap { workspaceContext =>
+      dataSource.inTransaction { dataAccess =>
+        withEntity(workspaceContext, entityType, entityName, dataAccess) { entity =>
+          dataAccess.entityQuery.get(workspaceContext, entity.entityType, newName) flatMap {
+            case None => dataAccess.entityQuery.rename(workspaceContext, entity.entityType, entity.name, newName)
+            case Some(_) => throw new RawlsExceptionWithErrorReport(errorReport = ErrorReport(StatusCodes.Conflict, s"Destination ${entity.entityType} ${newName} already exists"))
+          } map(_ => RequestComplete(StatusCodes.NoContent))
+        }
+      }
+    }
+
+  def evaluateExpression(workspaceName: WorkspaceName, entityType: String, entityName: String, expression: String): Future[PerRequestMessage] =
+    getWorkspaceContextAndPermissions(workspaceName, SamWorkspaceActions.read) flatMap { workspaceContext =>
+      dataSource.inTransaction { dataAccess =>
+        withSingleEntityRec(entityType, entityName, workspaceContext, dataAccess) { entities =>
+          ExpressionEvaluator.withNewExpressionEvaluator(dataAccess, Some(entities)) { evaluator =>
+            evaluator.evalFinalAttribute(workspaceContext, expression).asTry map { tryValuesByEntity => tryValuesByEntity match {
+              //parsing failure
+              case Failure(regret) => throw new RawlsExceptionWithErrorReport(errorReport = ErrorReport(StatusCodes.BadRequest, regret))
+              case Success(valuesByEntity) =>
+                if (valuesByEntity.size != 1) {
+                  //wrong number of entities?!
+                  throw new RawlsException(s"Expression parsing should have returned a single entity for ${entityType}/$entityName $expression, but returned ${valuesByEntity.size} entities instead")
+                } else {
+                  assert(valuesByEntity.head._1 == entityName)
+                  valuesByEntity.head match {
+                    case (_, Success(result)) => RequestComplete(StatusCodes.OK, result.toSeq)
+                    case (_, Failure(regret)) =>
+                      throw new RawlsExceptionWithErrorReport(errorReport = ErrorReport(StatusCodes.BadRequest, "Unable to evaluate expression '${expression}' on ${entityType}/${entityName} in ${workspaceName}", ErrorReport(regret)))
+                  }
+                }
+            }
+            }
+          }
+        }
+      }
+    }
+
+  def entityTypeMetadata(workspaceName: WorkspaceName): Future[PerRequestMessage] =
+    getWorkspaceContextAndPermissions(workspaceName, SamWorkspaceActions.read) flatMap { workspaceContext =>
+      dataSource.inTransaction { dataAccess =>
+        dataAccess.entityQuery.getEntityTypeMetadata(workspaceContext).map(r => RequestComplete(StatusCodes.OK, r))
+      }
+    }
+
+  def listEntities(workspaceName: WorkspaceName, entityType: String): Future[PerRequestMessage] =
+    getWorkspaceContextAndPermissions(workspaceName, SamWorkspaceActions.read) flatMap { workspaceContext =>
+      dataSource.inTransaction { dataAccess =>
+        dataAccess.entityQuery.listActiveEntitiesOfType(workspaceContext, entityType).map(r => RequestComplete(StatusCodes.OK, r.toSeq))
+      }
+    }
+
+  def queryEntities(workspaceName: WorkspaceName, entityType: String, query: EntityQuery): Future[PerRequestMessage] = {
+    getWorkspaceContextAndPermissions(workspaceName, SamWorkspaceActions.read) flatMap { workspaceContext =>
+      dataSource.inTransaction { dataAccess =>
+        dataAccess.entityQuery.loadEntityPage(workspaceContext, entityType, query) map { case (unfilteredCount, filteredCount, entities) =>
+          createEntityQueryResponse(query, unfilteredCount, filteredCount, entities.toSeq).get
+        }
+      }
+    }
+  }
+
+  def createEntityQueryResponse(query: EntityQuery, unfilteredCount: Int, filteredCount: Int, page: Seq[Entity]): Try[RequestComplete[(StatusCodes.Success, EntityQueryResponse)]] = {
+    val pageCount = Math.ceil(filteredCount.toFloat / query.pageSize).toInt
+    if (filteredCount > 0 && query.page > pageCount) {
+      Failure(new RawlsExceptionWithErrorReport(ErrorReport(StatusCodes.BadRequest, s"requested page ${query.page} is greater than the number of pages $pageCount")))
+
+    } else {
+      val response = EntityQueryResponse(query, EntityQueryResultMetadata(unfilteredCount, filteredCount, pageCount), page)
+
+      Success(RequestComplete(StatusCodes.OK, response))
+    }
+  }
+
+  def copyEntities(entityCopyDef: EntityCopyDefinition, uri: Uri, linkExistingEntities: Boolean): Future[PerRequestMessage] =
+
+    getWorkspaceContextAndPermissions(entityCopyDef.destinationWorkspace, SamWorkspaceActions.write) flatMap { destWorkspaceContext =>
+      getWorkspaceContextAndPermissions(entityCopyDef.sourceWorkspace,SamWorkspaceActions.read) flatMap { sourceWorkspaceContext =>
+        dataSource.inTransaction { dataAccess =>
+          for {
+            sourceAD <- DBIO.from(samDAO.getResourceAuthDomain(SamResourceTypeNames.workspace, sourceWorkspaceContext.workspace.workspaceId, userInfo))
+            destAD <- DBIO.from(samDAO.getResourceAuthDomain(SamResourceTypeNames.workspace, destWorkspaceContext.workspace.workspaceId, userInfo))
+            result <- authDomainCheck(sourceAD.toSet, destAD.toSet) flatMap { _ =>
+              val entityNames = entityCopyDef.entityNames
+              val entityType = entityCopyDef.entityType
+              val copyResults = dataAccess.entityQuery.checkAndCopyEntities(sourceWorkspaceContext, destWorkspaceContext, entityType, entityNames, linkExistingEntities)
+              copyResults.map { response =>
+                if (response.hardConflicts.isEmpty && (response.softConflicts.isEmpty || linkExistingEntities)) RequestComplete(StatusCodes.Created, response)
+                else RequestComplete(StatusCodes.Conflict, response)
+              }
+            }
+          } yield result
+        }
+      }
+    }
+
+  def batchUpdateEntities(workspaceName: WorkspaceName, entityUpdates: Seq[EntityUpdateDefinition], upsert: Boolean = false): Future[PerRequestMessage] = {
+    val namesToCheck = for {
+      update <- entityUpdates
+      operation <- update.operations
+    } yield operation.name
+
+    withAttributeNamespaceCheck(namesToCheck) {
+      getWorkspaceContextAndPermissions(workspaceName, SamWorkspaceActions.write) flatMap { workspaceContext =>
+        dataSource.inTransaction { dataAccess =>
+          val updateTrialsAction = dataAccess.entityQuery.getActiveEntities(workspaceContext, entityUpdates.map(eu => AttributeEntityReference(eu.entityType, eu.name))) map { entities =>
+            val entitiesByName = entities.map(e => (e.entityType, e.name) -> e).toMap
+            entityUpdates.map { entityUpdate =>
+              entityUpdate -> (entitiesByName.get((entityUpdate.entityType, entityUpdate.name)) match {
+                case Some(e) =>
+                  Try(applyOperationsToEntity(e, entityUpdate.operations))
+                case None =>
+                  if (upsert) {
+                    Try(applyOperationsToEntity(Entity(entityUpdate.name, entityUpdate.entityType, Map.empty), entityUpdate.operations))
+                  } else {
+                    Failure(new RuntimeException("Entity does not exist"))
+                  }
+              })
+            }
+          }
+
+          val saveAction = updateTrialsAction flatMap { updateTrials =>
+            val errorReports = updateTrials.collect { case (entityUpdate, Failure(regrets)) =>
+              ErrorReport(s"Could not update ${entityUpdate.entityType} ${entityUpdate.name}", ErrorReport(regrets))
+            }
+            if (!errorReports.isEmpty) {
+              DBIO.failed(new RawlsExceptionWithErrorReport(ErrorReport(StatusCodes.BadRequest, "Some entities could not be updated.", errorReports)))
+            } else {
+              val t = updateTrials.collect { case (entityUpdate, Success(entity)) => entity }
+
+              dataAccess.entityQuery.save(workspaceContext, t)
+            }
+          }
+
+          saveAction.map(_ => RequestComplete(StatusCodes.NoContent))
+        }
+      }
+    }
+  }
+
+  /**
+    * Applies the sequence of operations in order to the entity.
+    *
+    * @param entity to update
+    * @param operations sequence of operations
+    * @throws AttributeNotFoundException when removing from a list attribute that does not exist
+    * @throws AttributeUpdateOperationException when adding or removing from an attribute that is not a list
+    * @return the updated entity
+    */
+  def applyOperationsToEntity(entity: Entity, operations: Seq[AttributeUpdateOperation]): Entity = {
+    entity.copy(attributes = applyAttributeUpdateOperations(entity, operations))
+  }
+
+  private def withEntity[T](workspaceContext: SlickWorkspaceContext, entityType: String, entityName: String, dataAccess: DataAccess)(op: (Entity) => ReadWriteAction[T]): ReadWriteAction[T] = {
+    dataAccess.entityQuery.get(workspaceContext, entityType, entityName) flatMap {
+      case None => DBIO.failed(new RawlsExceptionWithErrorReport(errorReport = ErrorReport(StatusCodes.NotFound, s"${entityType} ${entityName} does not exist in ${workspaceContext.workspace.toWorkspaceName}")))
+      case Some(entity) => op(entity)
+    }
+  }
+
+}
